requirements.txt
bs4
sqlalchemy
requests
feedparser
dotenv
psycop2-binary
tqdm
google.generativeai